--- a/formats/cbor/commonTest/src/kotlinx/serialization/cbor/CborReaderTest.kt
+++ b/formats/cbor/commonTest/src/kotlinx/serialization/cbor/CborReaderTest.kt
@@ -11,7 +11,7 @@ import kotlin.test.*
 class CborReaderTest {
     private fun withDecoder(input: String, block: Cbor.CborDecoder.() -> Unit) {
         val bytes = HexConverter.parseHexBinary(input.toUpperCase())
-        Cbor.CborDecoder(ByteArrayInputStream(bytes)).block()
+        Cbor.CborDecoder(ByteArrayInput(bytes)).block()
     }
 
     @Test