--- a/formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/ProtoTypes.kt
+++ b/formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/ProtoTypes.kt
@@ -20,14 +20,10 @@ private const val MASK = Int.MAX_VALUE.toLong() shl 32
  * See [https://developers.google.com/protocol-buffers/docs/proto#scalar]
  */
 @Suppress("NO_EXPLICIT_VISIBILITY_IN_API_MODE_WARNING")
-public enum class ProtoNumberType(@JvmField internal val signature: Long) {
-    DEFAULT(1L shl 32),
-    SIGNED(2L shl 32),
-    FIXED(3L shl 32);
-
-    internal fun equalTo(descriptor: ProtoDesc): Boolean {
-        return descriptor and MASK == signature
-    }
+public enum class ProtoNumberType(internal val signature: Long) {
+    DEFAULT(0L shl 32),
+    SIGNED(1L shl 32),
+    FIXED(2L shl 32);
 }
 
 /**