… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/ProtobufReader.kt
… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/ProtobufTaggedDecoder.kt
… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/ProtobufTaggedEncoder.kt
… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/ProtobufWriter.kt
… formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/Varint.kt
… formats/protobuf/commonTest/src/kotlinx/serialization/protobuf/ProtobufMissingFieldsTest.kt
… formats/protobuf/commonTest/src/kotlinx/serialization/protobuf/ProtobufPrimitiveWrappersTest.kt
… formats/protobuf/commonTest/src/kotlinx/serialization/protobuf/ProtobufPrimitivesTest.kt
… formats/protobuf/jsMain/src/kotlinx/serialization/protobuf/Bytes.kt
… formats/protobuf/jvmMain/src/kotlinx/serialization/protobuf/Bytes.kt
… formats/protobuf/jvmTest/src/kotlinx/serialization/protobuf/ProtobufTopLevelPrimitivesCompatibilityTest.kt
… formats/protobuf/nativeMain/src/kotlinx/serialization/protobuf/Bytes.kt
… formats/protobuf/nativeTest/src/kotlinx/serialization/NativeTest.kt
… runtime/commonMain/src/kotlinx/io/Streams.common.kt
… runtime/commonMain/src/kotlinx/io/Streams.kt
… runtime/commonMain/src/kotlinx/serialization/Encoding.kt
… runtime/commonMain/src/kotlinx/serialization/internal/Tagged.kt
… runtime/commonMain/src/kotlinx/serialization/internal/Util.kt
… runtime/jsMain/src/kotlinx/io/Streams.kt
… runtime/jvmMain/src/kotlinx/io/Buffers.kt
… runtime/jvmMain/src/kotlinx/io/Streams.kt
… runtime/nativeMain/src/kotlinx/io/Streams.kt